services/automation-miner/requirements-dev.txt
# Development / Test dependencies
-r requirements.txt

pytest>=8.3.3
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-httpx>=0.30.0
